requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
requests==2.32.3
pydantic==2.7.0
python-dotenv==1.0.1
pgvector==0.3.2
pyyaml==6.0.2
